Add core.create/edit/delete ACL actions, separate from view access (#203)

+++ admin/src/Administrator/View/Allowlist/HtmlView.php
@@ -11,6 +11,7 @@ namespace Codeling\Component\Bfstop\Administrator\View\Allowlist;
 defined('_JEXEC') or die;
 
 use Codeling\Component\Bfstop\Administrator\Helper\ToolbarHelper as BfstopToolbarHelper;
+use Joomla\CMS\Factory;
 use Joomla\CMS\Language\Text;
 use Joomla\CMS\MVC\View\HtmlView as BaseHtmlView;
 use Joomla\CMS\Toolbar\ToolbarHelper;
@@ -30,11 +31,21 @@ class HtmlView extends BaseHtmlView
 
 	protected function addToolBar()
 	{
+		$user = Factory::getApplication()->getIdentity();
 		ToolbarHelper::title(Text::_('COM_BFSTOP_HEADING_ALLOWLIST'), 'bfstop');
 		ToolbarHelper::divider();
-		ToolbarHelper::deleteList('COM_BFSTOP_ALLOWLIST_DELETE_CONFIRM', 'allowlist.remove');
-		ToolbarHelper::editList('allow.edit');
-		ToolbarHelper::addNew('allow.add');
+		if ($user->authorise('core.delete', 'com_bfstop'))
+		{
+			ToolbarHelper::deleteList('COM_BFSTOP_ALLOWLIST_DELETE_CONFIRM', 'allowlist.remove');
+		}
+		if ($user->authorise('core.edit', 'com_bfstop'))
+		{
+			ToolbarHelper::editList('allow.edit');
+		}
+		if ($user->authorise('core.create', 'com_bfstop'))
+		{
+			ToolbarHelper::addNew('allow.add');
+		}
 		BfstopToolbarHelper::addOptions();
 	}
 }
